+++ src/LiGet/Cache/Catalog/ICatalogScanStore.cs
@@ -1,0 +1,9 @@
+using System;
+
+namespace LiGet.Cache.Catalog
+{
+    public interface ICatalogScanStore
+    {
+        DateTimeOffset LastScanEndDate { get; set; }
+    }
+}