+++ src/LiGet/Cache/Catalog/ICatalogScanner.cs
@@ -1,0 +1,9 @@
+using System;
+
+namespace LiGet.Cache.Catalog
+{
+    public interface ICatalogScanner
+    {
+        event EventHandler<CatalogEntryEventArgs> UpdatedEntry;
+    }
+}