+++ src/LiGet/Cache/DBreeze/IDBreezeEngine.cs
@@ -1,0 +1,7 @@
+namespace LiGet.Cache.DBreeze
+{
+    public interface IDBreezeEngineProvider
+    {
+         global::DBreeze.DBreezeEngine Engine { get; }
+    }
+}